% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.landscape.windows.random.R
\name{landscape.windows.random}
\alias{landscape.windows.random}
\title{Landscape windows: random method}
\usage{
landscape.windows.random(data, shape.path, proj, windows, scales,
  min.no.sites = 25, area = 80, nrowcol = 3, plot = FALSE)
}
\arguments{
\item{data}{Informations about sites (spatial coordinates). 'data' must have 3 columns: site ID, x and y.}

\item{shape.path}{Absolute path to the shapefile (study area).}

\item{proj}{Projection system. 'proj' must be expressed in EPSG code or in CRS.}

\item{windows}{Informations about the desired number of windows. 'windows' must contain 2 values: the desired number of windows per scale and the number of attempts (should be a multiplicative factor to multiply the desired number of windows).}

\item{scales}{Spatial scales to test. 'scales' must contain 3 values: the minimum windows' size, the maximum windows' size and the gap/break between two scales (in kilometers).}

\item{min.no.sites}{Minimum number of sites in each window (default = 25).}

\item{area}{Minimum percentage for the sites distributed over the window’s area to ensure equal sampling effort across categories (default = 80).}

\item{nrowcol}{Area is calculated with a grid of n rows/cols. 'nrowcol' determines n value (default = 3).}

\item{plot}{TRUE or FALSE. Creation of a map of the study area with the centers of the validated windows (default = FALSE).}
}
\value{
Output data is a list of 3 objects: a table containing all the validated windows, for each scale, with a list of sites present in each window; a SpatialPointsDataFrame containing the data (sites); and a SpatialPolygon of the study area.
}
\description{
This function creates landscape windows into an area defined by a shapefile (shape format) using the 'random method'. n windows are placed randomly in the defined area. With default values, to be validated, a window must contain at least 25 sites distributed over at least 80 percents of the window’s area (at least 1 site in 8 cells of a 3 x 3 grid).
}
